+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## HEAD
 
+Breaking changes
+
+- `x/txfee` extension decorator is a no-op wrapper when configuration for
+  `x/txfee` does not exist.
 
 ## 0.25.0